requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langgraph
faiss-cpu
sentence-transformers
streamlit
python-dotenv
beautifulsoup4
pygithub
markdown
duckduckgo-search